+++ libs/ymir-core/include/ymir/sys/saturn.hpp
@@ -14,6 +14,9 @@ See @ref index for instructions on how to use the emulator.
 
 #include <ymir/state/state.hpp>
 
+#include <ymir/debug/debug_callbacks.hpp>
+#include <ymir/debug/debug_internal_callbacks.hpp>
+
 #include "memory.hpp"
 #include "system.hpp"
 #include "system_features.hpp"
@@ -379,6 +382,29 @@ private:
     };
 
     SMPCOperations smpcOps{*this};
+
+    // -------------------------------------------------------------------------
+    // Debugger
+
+public:
+    /// @brief Sets the debug break callback to be invoked when the debug break signal is raised.
+    /// @param[in] callback the debug break callback
+    void SetDebugBreakRaisedCallback(debug::CBDebugBreakRaised callback) {
+        m_cbDebugBreakRaised = callback;
+    }
+
+private:
+    /// @brief Signals a debug break which interrupts emulation and invokes the attached debug break handler.
+    void SignalDebugBreak(/* TODO: const DebugBreakInfo &info */);
+
+    bool m_debugBreak = false; ///< Debug break signal
+
+    /// @brief Callback invoked when the debug break signal is raised.
+    debug::CBDebugBreakRaised m_cbDebugBreakRaised;
+
+public:
+    /// @brief The debug break signal callback to be used by other components.
+    debug::CBRaiseDebugBreak CbRaiseDebugBreak = util::MakeClassMemberRequiredCallback<&Saturn::SignalDebugBreak>(this);
 };
 
 } // namespace ymir